+++ OtripleS.Portal.Web.Tests.Unit/Services/Students/StudentServiceTests.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Net.Http;
 using Moq;
-using OtripleS.Portal.Web.Brokers.API;
+using OtripleS.Portal.Web.Brokers.Apis;
 using OtripleS.Portal.Web.Brokers.Logging;
 using OtripleS.Portal.Web.Models.Students;
 using OtripleS.Portal.Web.Services.Students;
