+++ OtripleS.Portal.Web/Brokers/Apis/ApiBroker.Classrooms.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using OtripleS.Portal.Web.Models.Classrooms;
 
-namespace OtripleS.Portal.Web.Brokers.API
+namespace OtripleS.Portal.Web.Brokers.Apis
 {
     public partial class ApiBroker
     {
