+++ OtripleS.Portal.Web/Brokers/Apis/IApiBroker.Teachers.cs
@@ -3,10 +3,14 @@
 // FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
 // ---------------------------------------------------------------
 
-namespace OtripleS.Portal.Web.Brokers.API
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using OtripleS.Portal.Web.Models.Teachers;
+
+namespace OtripleS.Portal.Web.Brokers.Apis
 {
     public partial interface IApiBroker
     {
-
+        ValueTask<List<Teacher>> GetAllTeachersAsync();
     }
 }
