+++ OtripleS.Portal.Web/Brokers/Apis/IApiBroker.cs
@@ -3,14 +3,10 @@
 // FREE TO USE AS LONG AS SOFTWARE FUNDS ARE DONATED TO THE POOR
 // ---------------------------------------------------------------
 
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using OtripleS.Portal.Web.Models.Courses;
-
-namespace OtripleS.Portal.Web.Brokers.API
+namespace OtripleS.Portal.Web.Brokers.Apis
 {
     public partial interface IApiBroker
     {
-        ValueTask<List<Course>> GetAllCoursesAsync();
+
     }
 }
